+++ src/lib.rs
@@ -13,10 +13,7 @@ mod axis;
 pub use axis::Axis;
 
 mod hat;
-pub use hat::Hat;
+pub use hat::{Hat, HatState};
 
 mod button;
 pub use button::{Button, ButtonState};
-
-const AXES_DISPLAY_NAMES: [&str; 8] = ["X", "Y", "Z", "Rx", "Ry", "Rz", "Slider", "Dial/Slider2"];
-const AXES_HID_USAGE: [u32; 8] = [0x30, 0x31, 0x32, 0x33, 0x34, 0x35, 0x36, 0x37];