+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## [0.2.0] Hat switch support
+- Added: Support for up to four 4-way hat switches (e.g. a D-pad).
 
 ## [0.1.0] Initial release